Refactor the code and reogranize packages
Reorganize the repository structure for better modularity and
maintainence. Refactor code to have proper references.
