+++ main.go
@@ -2,16 +2,11 @@ package main
 
 import (
 	"flag"
-	//"github.com/go-redis/redis"
 	"log"
-	//"go.uber.org/zap"
-	//"time"
-	//"strconv"
-	"context"
-	"github.com/aniketalshi/go_rest_cache/config"
-	"github.com/aniketalshi/go_rest_cache/logging"
-	"github.com/aniketalshi/go_rest_cache/db"
 	"net/http"
+
+	"github.com/aniketalshi/go_rest_cache/config"
+	"github.com/aniketalshi/go_rest_cache/app"
 )
 
 func main() {
@@ -25,30 +20,13 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-	
-	// initialize the logger
-	logging.InitLogger()
-	
-	cacher := &Cacher {
-		gitClient: GetNewGithubClient(context.Background()),
-		dbClient: db.SetupDBClient(),
-	}
 
-	contexedHandler := SetupHandlers(cacher)
+	// initialize the application
+	app := &app.App{}
+	app.Initialize()
+	app.Run()
 	
-	// used for synchronizing two different go routines so that one can 
-	// let the other know when data is cached
-	isCached := make(chan bool)
-	
-	go cacher.cache_repos(isCached)
-	go cacher.cache_members()
-	go cacher.cache_org_details()
-	go cacher.cache_root_endpoint()
-	
-	//time.Sleep(10 * time.Second)
-	go cacher.populate_views(isCached)
-
-	log.Fatal(http.ListenAndServe(":" + *httpPort, contexedHandler))
-	
+	// launch the server 
+	log.Fatal(http.ListenAndServe(":" + *httpPort, app.Handler))
 }
